+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- You can now use `$smarty->setPHP7CompatMode()` to activate php7 compatibility mode when running PHP8
+- You can now use `$smarty->muteUndefinedOrNullWarnings()` to activate convert warnings about undefined or null template vars to notices when running PHP8
 
 ### Changed
 - Switch CI from Travis to Github CI